+++ DemoApps/GRDBDemoiOS/GRDBDemoiOS/Person.swift
@@ -29,7 +29,7 @@ class Person : Record {
         
     }
     
-    override var storedDatabaseDictionary: [String: DatabaseValueConvertible?] {
+    override var persistedDictionary: [String: DatabaseValueConvertible?] {
         return [
             "id": id,
             "firstName": firstName,