+++ GRDB Tests/Public/Record/RecordAwakeFromFetchTests.swift
@@ -25,7 +25,7 @@ class EventRecorder : Record {
         super.init(row)
     }
     
-    override var storedDatabaseDictionary: [String: DatabaseValueConvertible?] {
+    override var persistedDictionary: [String: DatabaseValueConvertible?] {
         return ["id": id]
     }